init: spring data jpa, swagger 셋팅

+++ src/main/java/com/app/boilerplate/mapper/UserMapper.java
@@ -8,9 +8,9 @@ import org.apache.ibatis.annotations.Param;
 
 @Mapper
 public interface UserMapper {
-  UserResponse getUserById(int id);
+  UserResponse getUserById(Long id);
   List<UserResponse> getAllUser();
   void insertUser(UserRequest userRequest);
-  void updateUser(@Param("id") int id, @Param("userRequest") UserRequest userRequest);
-  void deleteUser(int id);
+  void updateUser(@Param("id") Long id, @Param("userRequest") UserRequest userRequest);
+  void deleteUser(Long id);
 }
